Adding create_yomidic
Using the novels available this script generates a dictionary that can
be imported into yomichan. Currently the terms will be ordered by
rankign e.g. Freq:1 means it's the term with the highest occurence.

It currently counts the surface fugashi created and the lemma. I'm not
sure what advantage this actually gives or rather what drawbacks it
creates.