Fixed a bug regarding furigana in some novels
Some of my novels do contain furigana inside a font tag with size 1. I
added a line to the novelmaker.py that filters them direclty with a
regex to ensure better quality in the final html and the tagging.
ex. 漁るhad the beginning kanji with asa furigana and due to how bs4
introduces newlines after every tag this resulted in 漁 あさ る in the
final html. Fugashi can't properly tag it and it reduces reading
experience as yomichan is also unable to parse it.