Change how the marking in novelmaker works
The marking process was notoriusly slow, which isn't such a huge deal
when it's only needed once a week but it was obvious that it was slow
due to my poor programming skills.

For simple string replacements I somehow decided to use re.sub instead
of string.replace which is a lot slower and that really adds when doing
the same operation several thousand times.

I know replaced all of that with a regex pattern builder und one simple
replacement. All of this results in quite the significant speedup from
around 1 minute per novel to 4ish seconds.