Update libraryreport.py to ingore full katawords
This is done due to most names being translated into katakana and most
words consisting of pure katakana are either sounds or foreign words for
whom i will decide while watching the content, if i want to mine them
but including them in the report would just clutter it. Names are
currently irrelevant for me which is why i'm ignoring them with this
change.